Create functionalty for extension points - Changed way of creating files for extension points

+++ Generator/createCustomFiles.py
@@ -7,19 +7,18 @@ constants = globals.Constants()
 
 def create_custom_functions(parsed_model):
     path = os.path.dirname(os.path.abspath(__file__))
-    model_entities = []
-    template_filename = "customs.py.jinja2"
+    template_filename = "model_extensions.py.jinja2"
     for model in parsed_model["modelItems"]:
         model_name = model["modelName"]
-        model_entities.append(model_name)
-    context = {
-        'm_models': model_entities
-    }
-    template_environment = Environment(
-        autoescape=False,
-        loader=FileSystemLoader(os.path.join(path, constants.goBack, constants.djangoCoreSource + "app\\")),
-        trim_blocks=False)
-    filename = constants.goBack + constants.targetDestination + "app\\" + "customs.py"
-    with open(filename, 'w') as f:
-        urls_py = template_environment.get_template(template_filename).render(context)
-        f.write(urls_py)
+        if(not os.path.exists(constants.goBack + constants.targetDestination + "app\\model_extensions\\" + model_name + "_extensions.py")):
+            context = {
+                'that_modelName': model_name
+            }
+            template_environment = Environment(
+                autoescape=False,
+                loader=FileSystemLoader(os.path.join(path, constants.goBack, constants.djangoCoreSource + "app\\model_extensions\\")),
+                trim_blocks=False)
+            filename = constants.goBack + constants.targetDestination + "app\\model_extensions\\" + model_name + "_extensions.py"
+            with open(filename, 'w') as f:
+                urls_py = template_environment.get_template(template_filename).render(context)
+                f.write(urls_py)
